requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
iso8601==1.1.0
lxml==4.9.1
multidict==6.0.2
pypika-tortoise==0.1.6
pytz==2022.2.1
soupsieve==2.3.2.post1
tortoise-orm==0.19.2
typing_extensions==4.3.0
yarl==1.8.1
